Add an option to specify the reception date of purchase order items
This enables logging reception of line items which happened in the past.
The date when the stock has been received is also helpful for better stock
value calculation in the future as the exchange rate depends on the
invoice date.

This can potentially break the stock ledger as the assumption that
stock item entries / stock history entries have increasing ids with
increasing dates is broken. More investigation and testing is needed.